+++ src/shared/csrf-xhr.ts
@@ -110,7 +110,8 @@ export class BBCsrfXhr {
     disableRedirect?: boolean,
     envId?: string,
     permissionScope?: string,
-    leId?: string
+    leId?: string,
+    bypassCsrf?: boolean
   ) {
     if (permissionScope && !envId) {
       return Promise.reject({
@@ -121,7 +122,18 @@ export class BBCsrfXhr {
 
     return new Promise((resolve: any, reject: any) => {
       // First get the CSRF token
-      requestToken(CSRF_URL, 'token_needed')
+
+      new Promise((resolveCsrf: any, rejectCsrf: any) => {
+        if (bypassCsrf) {
+          resolveCsrf({
+            csrf_token: 'token_needed'
+          });
+        } else {
+          requestToken(CSRF_URL, 'token_needed')
+            .then(resolveCsrf)
+            .catch(rejectCsrf);
+        }
+      })
         .then((csrfResponse: any) => {
           // Next get the access token, and then pass it to the callback.
           return requestToken(url, csrfResponse['csrf_token'], envId, permissionScope);
@@ -141,6 +153,35 @@ export class BBCsrfXhr {
             }
           }
         });
+    });
+  }
+
+  public static requestWithToken(
+    url: string,
+    token: string
+  ) {
+    return new Promise((resolve, reject) => {
+      const xhr = new XMLHttpRequest();
+
+      xhr.onreadystatechange = () => {
+        if (xhr.readyState === 4) {
+          switch (xhr.status) {
+            case 200:
+              resolve(JSON.parse(xhr.responseText));
+              break;
+            default:
+              reject();
+              break;
+          }
+        }
+      };
+
+      xhr.open('GET', url, true);
+
+      xhr.setRequestHeader('Authorization', 'Bearer ' + token);
+      xhr.setRequestHeader('Accept', 'application/json');
+
+      xhr.send();
     });
   }
 }